requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
protobuf==3.6.1
pytz==2019.3
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.1
